Move as_classifier into carl.ml
